+++ src/main/java/domain/Participants/Participants.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.Participants;
 
 import java.util.Collections;
 import java.util.List;
