+++ src/main/java/domain/ladder/Height.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.Objects;
