+++ src/main/java/domain/ladder/StepPoint.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public enum StepPoint {
     PRESENT(true),
