+++ src/main/java/utils/RandomStepGenerator.java
@@ -1,6 +1,6 @@
 package utils;
 
-import domain.StepPoint;
+import domain.ladder.StepPoint;
 import java.util.Random;
 
 public class RandomStepGenerator implements StepGenerator {
