+++ src/test/java/domain/LadderResultTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.ladder.Ladder;
+import domain.result.LadderResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
