+++ src/test/java/domain/LadderTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.ladder.Ladder;
+import domain.ladder.Line;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
