+++ src/test/java/domain/PositionTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.Participants.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
