+++ src/test/java/domain/PrizesTest.java
@@ -3,6 +3,8 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.Participants.Participants;
+import domain.result.Prizes;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,7 +14,8 @@ public class PrizesTest {
     @Test
     @DisplayName("실행 결과가 참가자 수와 다를 때 예외가 발생한다.")
     void notMatchPrizeExceptionTest() {
-        assertThatThrownBy(() -> new Prizes(List.of("정상글자", "정상", "정상값"), new Participants(List.of("a", "b", "c", "d"))))
+        assertThatThrownBy(
+                () -> new Prizes(List.of("정상글자", "정상", "정상값"), new Participants(List.of("a", "b", "c", "d"))))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 실행 결과의 개수는 참가자 수와 일치해야 합니다.");
     }
